XEN vNUMA with adaptable vCPUs
This is a big commit including the following changes:

+ The per-cpu page frame cache and the buddy system are partitionned by the number of physical NUMA nodes.
In other words, they both are now NUMA-aware. The kernel can know which page comes from which node
because it knows the memory range on each physical node. First, pseudo-physical frame numbers are converted
to machine frame numbers. The later are then used to compare with the memory ranges to locate the physical node.

+ The memory policy component is modified in a way that it can accept and provide to userspace applications
the virtual NUMA information. The later is mapped to the physical NUMA information when interacting with
the page allocation component.

+ A counter representing the current topology verison is introduced. We also define a new system call for
userspace applications to obtain the virtual NUMA information including the version counter. When the OS receives
from the hypervisor an interrupt notifying a change in topology, the version counter get increased.

Note: only machines with 2 NUMA nodes are supported with this commit.